Added npm scripts to ease development workflow
Usage:
* `npm install` to install both npm and bower packages
* `npm test` to do once off build and test
* `npm start` to gulp watch and start a web server watching the directory

I've found this to be a huge benefit when simplifying CI build
configurations. Something that isn't obvious is that bower and gulp
don't actually need to be on the path for these scripts to work. npm
will check for the ones installed during the npm install step.  This
also helps in getting new devs up and running, since there are less
global dependencies for them to install.